Put glfe stuff in a different file.
