+++ include/tools/glfw.hpp
@@ -1,0 +1,9 @@
+#include <GLFW/glfw3.h>
+
+void drawPoint(float x, float y, float radius, int segments, float red, float green, float blue);
+
+void drawLine(float x_initial, float y_initial, float x_final, float y_final, float red, float green, float blue);
+
+void drawRectangle(float x, float y, float width, float height, float red, float green, float blue);
+
+void drawGrid(int rows, int cols, int width, int height);
